% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{Arith,dcudaMatrix,dcudaMatrix-method}
\alias{Arith,dcudaMatrix,dcudaMatrix-method}
\title{Double Precision CUDA Matrix Addition/Subtraction}
\usage{
\S4method{Arith}{dcudaMatrix,dcudaMatrix}(e1, e2)
}
\arguments{
\item{e1}{An dcudaMatrix object}

\item{e2}{An dcudaMatrix object}
}
\description{
Double Precision CUDA Matrix Addition/Subtraction
}
