% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{Arith,fcudaMatrix,fcudaMatrix-method}
\alias{Arith,fcudaMatrix,fcudaMatrix-method}
\title{Single Precision CUDA Matrix Addition/Subtraction}
\usage{
\S4method{Arith}{fcudaMatrix,fcudaMatrix}(e1, e2)
}
\arguments{
\item{e1}{An fcudaMatrix object}

\item{e2}{An fcudaMatrix object}
}
\description{
Single Precision CUDA Matrix Addition/Subtraction
}
