% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/wrappers.R
\name{cudaMatMult}
\alias{cudaMatMult}
\title{CUDA GPU Matrix Multiplication}
\usage{
cudaMatMult(A, B)
}
\arguments{
\item{A}{A cudaMatrix object}

\item{B}{A cudaMatrix object}
}
\value{
A cudaMatrix object
}
\description{
matrix multiplication
}
\author{
Charles Determan Jr.
}

