% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{colSums,cudaMatrix,missing,missing-method}
\alias{colSums,cudaMatrix,missing,missing-method}
\alias{colSums,cudaMatrix}
\alias{rowSums,cudaMatrix,missing,missing-method}
\alias{colMeans,cudaMatrix,missing,missing-method}
\alias{rowMeans,cudaMatrix,missing,missing-method}
\title{Row and Column Sums and Means of cudaMatrix}
\usage{
\S4method{colSums}{cudaMatrix,missing,missing}(x, na.rm, dims)

\S4method{rowSums}{cudaMatrix,missing,missing}(x, na.rm, dims)

\S4method{colMeans}{cudaMatrix,missing,missing}(x, na.rm, dims)

\S4method{rowMeans}{cudaMatrix,missing,missing}(x, na.rm, dims)
}
\arguments{
\item{x}{A cudaMatrix object}

\item{na.rm}{Not currently used}

\item{dims}{Not currently used}
}
\value{
A cudaVector object
}
\description{
Row and column sums and of cudaMatrix objects
}
\author{
Charles Determan Jr.
}
