% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cudaVector.R
\docType{methods}
\name{cudaVector}
\alias{cudaVector}
\alias{cudaVector,vector,missing-method}
\alias{cudaVector,vector}
\alias{cudaVector,missing,ANY-method}
\alias{cudaVector,missing}
\title{Construct a cudaVector}
\usage{
cudaVector(data, length, type = NULL, ...)

\S4method{cudaVector}{vector,missing}(data, length = NULL, type = NULL)

\S4method{cudaVector}{missing,ANY}(data, length, type = NULL)
}
\arguments{
\item{data}{An object that is or can be converted to a 
\code{vector}}

\item{length}{A non-negative integer specifying the desired length.}

\item{type}{A character string specifying the type of cudaVector.  Default
is NULL where type is inherited from the source data type.}

\item{...}{Additional method to pass to cudaVector methods}
}
\value{
A cudaVector object
}
\description{
Construct a cudaVector of a class that inherits
from \code{cudaVector}
}
\author{
Charles Determan Jr.
}
