% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{dcudaMatrix-class}
\alias{dcudaMatrix-class}
\title{dcudaMatrix Class}
\description{
An integer type matrix in the S4 \code{cudaMatrix}
representation.
}
\section{Slots}{

 \describe{
     \item{\code{x}:}{A numeric R matrix}
     \item{\code{type}:}{Character object specifying
     the type the matrix data is double}
 }
}
\author{
Charles Determan Jr.
}
\seealso{
\code{\link{cudaMatrix-class}},
\code{\link{icudaMatrix-class}},
\code{\link{fcudaMatrix-class}}
}

