% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{dim,cudaMatrix-method}
\alias{dim,cudaMatrix}
\alias{dim,cudaMatrix-method}
\title{cudaMatrix dim method}
\usage{
\S4method{dim}{cudaMatrix}(x)
}
\arguments{
\item{x}{A cudaMatrix object}
}
\value{
A length 2 vector of the number of rows and columns respectively.
}
\description{
cudaMatrix dim method
}
\author{
Charles Determan Jr.
}

