% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{dim,gpuRcudaMatrix-method}
\alias{dim,gpuRcudaMatrix-method}
\alias{dim,cudaMatrix}
\title{gpuRcuda dim method}
\usage{
\S4method{dim}{gpuRcudaMatrix}(x)
}
\arguments{
\item{x}{A gpuRcuda matrix object}
}
\value{
A length 2 vector of the number of rows and columns respectively.
}
\author{
Charles Determan Jr.
}
