% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-nvVector.R
\docType{class}
\name{dnvVector-class}
\alias{dnvVector-class}
\title{dnvVector Class}
\description{
An double vector in the S4 \code{nvVector}
representation.
}
\section{Slots}{

 \describe{
     \item{\code{address}:}{Pointer to a double typed vector}
 }
}

\seealso{
\code{\link{nvVector-class}}
}
\author{
Charles Determan Jr.
}
