% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-cudaVector.R, R/methods-nvVector.R,
%   R/methods.R
\docType{methods}
\name{[,cudaVector,missing,missing,missing-method}
\alias{[,cudaVector,missing,missing,missing-method}
\alias{[,nvVector,missing,missing,missing-method}
\alias{[,cudaMatrix,missing,missing,missing-method}
\alias{[,nvMatrix,missing,missing,missing-method}
\title{Extract gpuRcuda elements}
\usage{
\S4method{[}{cudaVector,missing,missing,missing}(x, i, j, drop)

\S4method{[}{nvVector,missing,missing,missing}(x, i, j, drop)

\S4method{[}{cudaMatrix,missing,missing,missing}(x, i, j, drop)

\S4method{[}{nvMatrix,missing,missing,missing}(x, i, j, drop)
}
\arguments{
\item{x}{A gpuRcuda object}

\item{i}{missing}

\item{j}{missing}

\item{drop}{missing}
}
\author{
Charles Determan Jr.
}
