% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-cudaMatrix.R
\docType{class}
\name{f32cudaMatrix-class}
\alias{f32cudaMatrix-class}
\title{f32cudaMatrix Class}
\description{
An float32 type matrix in the S4 \code{cudaMatrix}
representation.
}
\section{Slots}{

 \describe{
     \item{\code{address}:}{A float typed R matrix}
 }
}

\seealso{
\code{\link{cudaMatrix-class}}, 
\code{\link{icudaMatrix-class}},
\code{\link{dcudaMatrix-class}}
}
\author{
Chaitanya Inumella.
}
