% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cudaVector-class.R
\docType{class}
\name{fcudaVector-class}
\alias{fcudaVector-class}
\title{fcudaVector Class}
\description{
An float vector in the S4 \code{cudaVector}
representation.
}
\section{Slots}{

 \describe{
     \item{\code{address}:}{Pointer to a float typed vector}
 }
}
\author{
Charles Determan Jr.
}
\seealso{
\code{\link{cudaVector-class}}
}

