% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cudaMatrix.R
\docType{methods}
\name{float32}
\alias{float32}
\title{float32 datatype generic function}
\usage{
float32(x, ...)
}
\arguments{
\item{...}{Additional method to pass to cudaMatrix methods}

\item{data}{An object that is or can be converted to a 
\code{datatype} object}

\item{ncol}{An integer specifying the number of columns}

\item{nrow}{An integer specifying the number of rows}

\item{type}{A character string specifying the type of cudaMatrix.  Default
is NULL where type is inherited from the source data type.}
}
\value{
A float32 object
}
\description{
Construct a cudaMatrix of a class that inherits
from \code{float32}
}
\author{
Chaitanya Inumella
}
