% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-nvVector.R
\docType{class}
\name{fnvVector-class}
\alias{fnvVector-class}
\title{fnvVector Class}
\description{
An float vector in the S4 \code{nvVector}
representation.
}
\section{Slots}{

 \describe{
     \item{\code{address}:}{Pointer to a float typed vector}
 }
}

\seealso{
\code{\link{nvVector-class}}
}
\author{
Charles Determan Jr.
}
