\name{gpuRcuda-package}
\alias{gpuRcuda-package}
\alias{gpuRcuda}
\docType{package}
\title{
\packageTitle{gpuRcuda}
}
\description{
\packageDescription{gpuRcuda}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{gpuRcuda}
\packageIndices{gpuRcuda}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{gpuRcuda}

Maintainer: \packageMaintainer{gpuRcuda}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
%% ~~ simple examples of the most important functions ~~
}
