% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\docType{class}
\name{gpuRcudaVector-class}
\alias{gpuRcudaVector-class}
\title{gpuRcudaVector Class}
\description{
This is the 'mother' class for all
gpuRcudaVector objects.  It is essentially a wrapper for
a basic R vector (possibly to be improved).  All other 
gpuRcuda classes inherit from this class but 
there are no current circumstances where this class 
is used directly.
}
\note{
R does not contain a native float type.  As such,
the vector data within a such an object 
will be represented as double but downcast when any 
gpuRcuda methods are used.

May also remove the type slot
}
\section{Slots}{

 Common to all gpuRcuda objects in the package
 \describe{
     \item{\code{address}:}{An R vector object}
 }
}

\seealso{
\code{\link{nvVector-class}}, 
\code{\link{cudaVector-class}}
}
\author{
Charles Determan Jr.
}
