% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{\%*\%,cudaMatrix,cudaMatrix-method}
\alias{\%*\%,cudaMatrix,cudaMatrix-method}
\title{CUDA Matrix Multiplication}
\usage{
\S4method{\%*\%}{cudaMatrix,cudaMatrix}(x, y)
}
\arguments{
\item{x}{An fcudaMatrix object}

\item{y}{An fcudaMatrix object}
}
\description{
CUDA Matrix Multiplication
}
