% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{\%*\%,fcudaMatrix,fcudaMatrix-method}
\alias{\%*\%,fcudaMatrix,fcudaMatrix-method}
\title{Single Precision CUDA Matrix Multiplication}
\usage{
\S4method{\%*\%}{fcudaMatrix,fcudaMatrix}(x, y)
}
\arguments{
\item{x}{An fcudaMatrix object}

\item{y}{An fcudaMatrix object}
}
\description{
Single Precision CUDA Matrix Multiplication
}
