% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-cudaMatrix.R
\docType{class}
\name{i64cudaMatrix-class}
\alias{i64cudaMatrix-class}
\title{i64cudaMatrix Class}
\description{
An integer64 type matrix in the S4 \code{cudaMatrix}
representation.
}
\section{Slots}{

 \describe{
     \item{\code{address}:}{A integer64 typed R matrix}
 }
}

\seealso{
\code{\link{cudaMatrix-class}}, 
\code{\link{i64cudaMatrix-class}},
\code{\link{dcudaMatrix-class}}
}
\author{
Chaitanya Inumella
}
