% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-nvMatrix.R
\docType{class}
\name{invMatrix-class}
\alias{invMatrix-class}
\title{invMatrix Class}
\description{
An integer type matrix in the S4 \code{nvMatrix}
representation.
}
\section{Slots}{

 \describe{
     \item{\code{address}:}{A integer typed R matrix}
 }
}

\seealso{
\code{\link{nvMatrix-class}}, 
\code{\link{invMatrix-class}},
\code{\link{dnvMatrix-class}}
}
\author{
Charles Determan Jr.
}
