% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-nvVector.R
\docType{class}
\name{invVector-class}
\alias{invVector-class}
\title{invVector Class}
\description{
An integer vector in the S4 \code{nvVector}
representation.
}
\section{Slots}{

 \describe{
     \item{\code{address}:}{An integer vector object}
 }
}

\seealso{
\code{\link{nvVector-class}}
}
\author{
Charles Determan Jr.
}
