% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-vector.R
\docType{methods}
\name{length,gpuRcudaVector-method}
\alias{length,gpuRcudaVector-method}
\alias{length,nvVector}
\alias{length,cudaVector}
\title{Length of gpuRcudaVector}
\usage{
\S4method{length}{gpuRcudaVector}(x)
}
\arguments{
\item{x}{A gpuRcudaVector object}
}
\description{
Length of gpuRcudaVector
}
\author{
Charles Determan Jr.
}
