% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{nrow,gpuRcudaMatrix-method}
\alias{nrow,gpuRcudaMatrix-method}
\alias{nrow,cudaMatrix}
\alias{ncol,gpuRcudaMatrix-method}
\title{The Number of Rows/Columns of a gpuRcudaMatrix}
\usage{
\S4method{nrow}{gpuRcudaMatrix}(x)

\S4method{ncol}{gpuRcudaMatrix}(x)
}
\arguments{
\item{x}{A gpuRcudaMatrix object}
}
\value{
An integer of length 1
}
\description{
\code{nrow} and \code{ncol} return the number of rows
or columns present in \code{x}.
}
\author{
Charles Determan Jr.
}
