% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{nrow,cudaMatrix-method}
\alias{ncol,cudaMatrix-method}
\alias{nrow,cudaMatrix}
\alias{nrow,cudaMatrix-method}
\title{The Number of Rows/Columns of a cudaMatrix}
\usage{
\S4method{nrow}{cudaMatrix}(x)

\S4method{ncol}{cudaMatrix}(x)
}
\arguments{
\item{x}{A cudaMatrix object}
}
\value{
An integer of length 1
}
\description{
The Number of Rows/Columns of a cudaMatrix
}
\author{
Charles Determan Jr.
}

