% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-nvMatrix.R
\docType{class}
\name{nvMatrix-class}
\alias{nvMatrix-class}
\title{nvMatrix Class}
\description{
This class refers to gpuRcuda matrix
objects that are on the host.

There are multiple child classes that correspond
to the particular data type contained.  These include
\code{invMatrix}, \code{fnvMatrix}, and 
\code{dnvMatrix} corresponding to integer, float, and
double data types respectively.
}
\section{Slots}{

 Common to all nvMatrix objects in the package
 \describe{
     \item{\code{address}:}{An R matrix object}
 }
}

\seealso{
\code{\link{invMatrix-class}}, 
\code{\link{fnvMatrix-class}},
\code{\link{dnvMatrix-class}}
}
\author{
Charles Determan Jr.
}
