% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nvMatrix.R
\docType{methods}
\name{nvMatrix}
\alias{nvMatrix}
\alias{nvMatrix,matrix-method}
\alias{nvMatrix,matrix}
\alias{nvMatrix,missing-method}
\alias{nvMatrix,missing}
\title{Construct a nvMatrix}
\usage{
nvMatrix(data = NA, ncol = NA, nrow = NA, type = NULL, ...)

\S4method{nvMatrix}{matrix}(data, type = NULL)

\S4method{nvMatrix}{missing}(data, ncol = NA, nrow = NA, type = NULL)
}
\arguments{
\item{data}{An object that is or can be converted to a 
\code{matrix} object}

\item{ncol}{An integer specifying the number of columns}

\item{nrow}{An integer specifying the number of rows}

\item{type}{A character string specifying the type of nvMatrix.  Default
is NULL where type is inherited from the source data type.}

\item{...}{Additional method to pass to nvMatrix methods}
}
\value{
A nvMatrix object
}
\description{
Construct a nvMatrix of a class that inherits
from \code{nvMatrix}
}
\author{
Charles Determan Jr.
}
