% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-nvVector.R
\docType{class}
\name{nvVector-class}
\alias{nvVector-class}
\title{nvVector Class}
\description{
This is the 'mother' class for all
nvVector objects.  All other nvVector classes
inherit from this class but there are no current
circumstances where this class is used directly.

There are multiple child classes that correspond
to the particular data type contained.  These include
\code{invVector}.
}
\seealso{
\code{\link{invVector-class}}
}
\author{
Charles Determan Jr.
}
