% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{[,cudaMatrix,missing,missing,missing-method}
\alias{[,cudaMatrix}
\alias{[,cudaMatrix,missing,missing,missing-method}
\title{Extract all cudaMatrix elements}
\usage{
\S4method{[}{cudaMatrix,missing,missing,missing}(x, i, j, drop)
}
\arguments{
\item{x}{A cudaMatrix object}

\item{i}{missing}

\item{j}{missing}

\item{drop}{missing}
}
\description{
Extract all cudaMatrix elements
}
\author{
Charles Determan Jr.
}

