% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/typeof.R
\docType{methods}
\name{typeof,cudaMatrix-method}
\alias{typeof,cudaMatrix}
\alias{typeof,cudaMatrix-method}
\title{Get cudaMatrix type}
\usage{
\S4method{typeof}{cudaMatrix}(x)
}
\arguments{
\item{x}{A cudaMatrix object}
}
\description{
Get cudaMatrix type
}

