% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/methods_vector.R
\docType{methods}
\name{typeof,cudaVector-method}
\alias{typeof,cudaVector}
\alias{typeof,cudaVector-method}
\title{Get cudaVector type}
\usage{
\S4method{typeof}{cudaVector}(x)
}
\arguments{
\item{x}{A cudaVector object}
}
\description{
Get cudaVector type
}

