% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-cudaMatrix.R
\docType{class}
\name{cudaMatrix-class}
\alias{cudaMatrix-class}
\title{cudaMatrix Class}
\description{
This class refers to gpuRcuda matrix
objects that are on the device.

There are multiple child classes that correspond
to the particular data type contained.  These include
\code{icudaMatrix}, \code{fcudaMatrix}, and 
\code{dcudaMatrix} corresponding to integer, float, and
double data types respectively.
}
\note{
R does not contain a native float type.  As such,
the matrix data within a \code{\link{fcudaMatrix-class}} 
will be represented as double but downcast when any 
cudaMatrix methods are used.

May also remove the type slot
}
\section{Slots}{

 Common to all cudaMatrix objects in the package
 \describe{
     \item{\code{address}:}{An R matrix object}
 }
}

\seealso{
\code{\link{icudaMatrix-class}}, 
\code{\link{fcudaMatrix-class}},
\code{\link{dcudaMatrix-class}}
}
\author{
Charles Determan Jr.
}
