% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-cudaVector.R
\docType{class}
\name{cudaVector-class}
\alias{cudaVector-class}
\title{cudaVector Class}
\description{
This is the 'mother' class for all
cudaVector objects.  All other cudaVector classes
inherit from this class but there are no current
circumstances where this class is used directly.

There are multiple child classes that correspond
to the particular data type contained.  These include
\code{icudaVector}.
}
\seealso{
\code{\link{icudaVector-class}}
}
\author{
Charles Determan Jr.
}
