% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-cudaMatrix.R
\docType{class}
\name{dcudaMatrix-class}
\alias{dcudaMatrix-class}
\title{dcudaMatrix Class}
\description{
An integer type matrix in the S4 \code{cudaMatrix}
representation.
}
\section{Slots}{

 \describe{
     \item{\code{address}:}{A numeric R matrix}
 }
}

\seealso{
\code{\link{cudaMatrix-class}}, 
\code{\link{icudaMatrix-class}},
\code{\link{fcudaMatrix-class}}
}
\author{
Charles Determan Jr.
}
