% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-cudaVector.R
\docType{class}
\name{dcudaVector-class}
\alias{dcudaVector-class}
\title{dcudaVector Class}
\description{
An double vector in the S4 \code{cudaVector}
representation.
}
\section{Slots}{

 \describe{
     \item{\code{address}:}{Pointer to a double typed vector}
 }
}

\seealso{
\code{\link{cudaVector-class}}
}
\author{
Charles Determan Jr.
}
