% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{icudaMatrix-class}
\alias{icudaMatrix-class}
\title{icudaMatrix Class}
\description{
An integer type matrix in the S4 \code{cudaMatrix}
representation.
}
\section{Slots}{

 \describe{
     \item{\code{x}:}{A integer typed R matrix}
     \item{\code{type}:}{Character object specifying
     the type the matrix data is integer}
 }
}
\author{
Charles Determan Jr.
}
\seealso{
\code{\link{cudaMatrix-class}},
\code{\link{icudaMatrix-class}},
\code{\link{dcudaMatrix-class}}
}

