% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-cudaVector.R
\docType{class}
\name{icudaVector-class}
\alias{icudaVector-class}
\title{icudaVector Class}
\description{
An integer vector in the S4 \code{cudaVector}
representation.
}
\section{Slots}{

 \describe{
     \item{\code{address}:}{An integer vector object}
 }
}

\seealso{
\code{\link{cudaVector-class}}
}
\author{
Charles Determan Jr.
}
