% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/typeof.R
\name{typeof,gpuRcudaMatrix-method}
\alias{typeof,gpuRcudaMatrix-method}
\alias{typeof,cudaMatrix}
\alias{typeof,nvMatrix}
\alias{typeof,gpuRcudaVector-method}
\alias{typeof,cudaVector}
\alias{typeof,nvVector}
\title{Get gpuRcudaMatrix type}
\usage{
\S4method{typeof}{gpuRcudaMatrix}(x)

\S4method{typeof}{gpuRcudaVector}(x)
}
\arguments{
\item{x}{A gpuRcudaVector object}
}
\description{
Get gpuRcudaMatrix type

Get gpuRcudaVector type
}
\author{
Charles Determan Jr.

Charles Determan Jr.
}
